app-engine-by-example/requirements.txt
Flask==2.0.3
google-cloud-datastore==2.4.0
google-auth==2.6.0
requests==2.27.1